Remove auth handling endpoints
