+++ functions/src/auth.ts
@@ -1,15 +1,14 @@
 import { NextFunction, Response, Request } from "express";
-import { APIKeyHandler } from "./APIKeyHandler";
 import {
   sendAndThrowShareLinkOrUnexpectedError,
   ShareLinkError,
   ShareLinkErrorCode,
 } from "./error-handling";
 import { firebaseApp } from "./init";
-import { verifyIdToken } from "./utils";
+import { APIFields } from "./utils";
 
+const API_KEY_COLLECTION = "apiKeys";
 const firestoreDb = firebaseApp.firestore();
-const apiKeyHandler = new APIKeyHandler(firestoreDb);
 
 /**
  * Middleware to verify that the request is authorized with an API key.
@@ -24,8 +23,7 @@ export function isAPIKeyAuthorized(
   next: NextFunction
 ) {
   const apiKey = (req.query.apiKey as string) || (req.body.apiKey as string);
-  return apiKeyHandler
-    .isValidKey(apiKey)
+  return isValidKey(apiKey)
     .then((isValid) => {
       if (isValid) {
         next();
@@ -39,21 +37,20 @@ export function isAPIKeyAuthorized(
 }
 
 /**
- * Middleware to verify that the request is authorized with a valid Firebase ID token.
+ * Verify given API key exists and is enabled.
  *
- * @param req Express request to check for authorization.
- * @param res Express response to interact with.
- * @param next Next function.
+ * @param apiKey The API key to verify.
+ * @returns True if the API key exists and is enabled, false otherwise.
  */
-export function isFirebaseAuthorized(
-  req: Request,
-  res: Response,
-  next: NextFunction
-) {
-  const idToken = req.get("Authorization")?.split("Bearer ")[1];
-  verifyIdToken(idToken)
-    .then(() => next())
-    .catch((error) => {
-      sendAndThrowShareLinkOrUnexpectedError(error, res);
-    });
+async function isValidKey(apiKey: string | undefined): Promise<boolean> {
+  if (!apiKey) return false;
+  const apiKeyDoc = await firestoreDb
+    .collection(API_KEY_COLLECTION)
+    .where(APIFields.API_KEY, "==", apiKey)
+    .get();
+  if (apiKeyDoc.empty || !apiKeyDoc.docs[0].data().enabled) {
+    return false;
+  } else {
+    return true;
+  }
 }
